# More capable inliner [ci: last-only]