--- a/project/ScriptCommands.scala
+++ b/project/ScriptCommands.scala
@@ -103,7 +103,7 @@ object ScriptCommands {
   }
 
   private[this] val enableOptimizer = Seq(
-    scalacOptions in Compile in ThisBuild ++= Seq("-opt:l:inline", "-opt-inline-from:scala/**")
+    scalacOptions in Compile in ThisBuild ++= Seq("-opt:l:inline", "-opt-inline-from:scala.**")
   )
 
   private[this] val noDocs = Seq(