--- a/src/compiler/scala/tools/nsc/backend/jvm/BackendReporting.scala
+++ b/src/compiler/scala/tools/nsc/backend/jvm/BackendReporting.scala
@@ -192,8 +192,9 @@ object BackendReporting {
       val reason = this match {
         case CalleeNotFinal(_, _, _, _) =>
           s"The method is not final and may be overridden."
-        case IllegalAccessInstruction(_, _, _, _, callsiteClass, instruction) =>
-          s"The callee $calleeMethodSig contains the instruction ${AsmUtils.textify(instruction)}" +
+        case IllegalAccessInstructions(_, _, _, _, callsiteClass, instructions) =>
+          val suffix = if (instructions.lengthCompare(1) > 0) "s" else ""
+          s"The callee $calleeMethodSig contains the instruction$suffix ${instructions.map(AsmUtils.textify).mkString(", ")}" +
             s"\nthat would cause an IllegalAccessError when inlined into class $callsiteClass."
 
         case IllegalAccessCheckFailed(_, _, _, _, callsiteClass, instruction, cause) =>
@@ -226,8 +227,8 @@ object BackendReporting {
     }
   }
   case class CalleeNotFinal(calleeDeclarationClass: InternalName, name: String, descriptor: String,  annotatedInline: Boolean) extends CannotInlineWarning
-  case class IllegalAccessInstruction(calleeDeclarationClass: InternalName, name: String, descriptor: String, annotatedInline: Boolean,
-                                      callsiteClass: InternalName, instruction: AbstractInsnNode) extends CannotInlineWarning
+  case class IllegalAccessInstructions(calleeDeclarationClass: InternalName, name: String, descriptor: String, annotatedInline: Boolean,
+                                       callsiteClass: InternalName, instructions: List[AbstractInsnNode]) extends CannotInlineWarning
   case class IllegalAccessCheckFailed(calleeDeclarationClass: InternalName, name: String, descriptor: String, annotatedInline: Boolean,
                                       callsiteClass: InternalName, instruction: AbstractInsnNode, cause: OptimizerWarning) extends CannotInlineWarning
   case class MethodWithHandlerCalledOnNonEmptyStack(calleeDeclarationClass: InternalName, name: String, descriptor: String, annotatedInline: Boolean,