--- a/src/compiler/scala/tools/nsc/backend/jvm/PostProcessor.scala
+++ b/src/compiler/scala/tools/nsc/backend/jvm/PostProcessor.scala
@@ -15,6 +15,7 @@ package backend.jvm
 
 import java.util.concurrent.ConcurrentHashMap
 
+import scala.collection.mutable
 import scala.reflect.internal.util.{NoPosition, Position, StringContextStripMarginOps}
 import scala.reflect.io.AbstractFile
 import scala.tools.asm.ClassWriter
@@ -125,7 +126,7 @@ abstract class PostProcessor extends PerRunInit {
       if (compilerSettings.optInlinerEnabled)
         inliner.runInlinerAndClosureOptimizer()
       else if (compilerSettings.optClosureInvocations)
-        closureOptimizer.rewriteClosureApplyInvocations(None)
+        closureOptimizer.rewriteClosureApplyInvocations(None, mutable.Map.empty)
     }
   }