--- a/src/compiler/scala/tools/nsc/backend/jvm/PostProcessorFrontendAccess.scala
+++ b/src/compiler/scala/tools/nsc/backend/jvm/PostProcessorFrontendAccess.scala
@@ -65,6 +65,7 @@ object PostProcessorFrontendAccess {
 
     def optAddToBytecodeRepository: Boolean
     def optBuildCallGraph: Boolean
+    def optUseAnalyzerCache: Boolean
 
     def optNone: Boolean
     def optLClasspath: Boolean
@@ -185,6 +186,7 @@ object PostProcessorFrontendAccess {
 
       val optAddToBytecodeRepository: Boolean = s.optAddToBytecodeRepository
       val optBuildCallGraph: Boolean = s.optBuildCallGraph
+      val optUseAnalyzerCache: Boolean = s.optUseAnalyzerCache
 
       val optNone: Boolean = s.optNone
       val optLClasspath: Boolean = s.optLClasspath