--- a/test/files/run/bcodeInlinerMixed/A_1.java
+++ b/test/files/run/bcodeInlinerMixed/A_1.java
@@ -1,3 +1,4 @@
 public class A_1 {
-  public static final int bar() { return 100; }
+  // non-trivial method to avoid it from being inlined
+  public static final int bar() { return Integer.valueOf("123") + Integer.valueOf("321"); }
 }