… src/compiler/scala/tools/nsc/backend/jvm/analysis/NullnessAnalyzer.scala
… src/compiler/scala/tools/nsc/backend/jvm/analysis/ProdConsAnalyzer.scala
… src/compiler/scala/tools/nsc/backend/jvm/analysis/TypeFlowAnalyzer.scala
… src/compiler/scala/tools/nsc/backend/jvm/opt/BoxUnbox.scala
… src/compiler/scala/tools/nsc/backend/jvm/opt/ByteCodeRepository.scala
… src/compiler/scala/tools/nsc/backend/jvm/opt/BytecodeUtils.scala
… src/compiler/scala/tools/nsc/backend/jvm/opt/CallGraph.scala
… src/compiler/scala/tools/nsc/backend/jvm/opt/ClosureOptimizer.scala
… src/compiler/scala/tools/nsc/backend/jvm/opt/CopyProp.scala
… src/compiler/scala/tools/nsc/backend/jvm/opt/Inliner.scala
… src/compiler/scala/tools/nsc/backend/jvm/opt/InlinerHeuristics.scala
… src/compiler/scala/tools/nsc/backend/jvm/opt/LocalOpt.scala
… src/compiler/scala/tools/nsc/settings/ScalaSettings.scala
… src/library/scala/Predef.scala
… src/library/scala/collection/ArrayOps.scala
… src/library/scala/collection/mutable/ArrayBuilder.scala
… src/library/scala/reflect/ClassTag.scala
… src/library/scala/reflect/Manifest.scala
… src/library/scala/runtime/LambdaDeserialize.java
… src/library/scala/runtime/LambdaDeserialize.scala
… src/library/scala/runtime/MethodHandleConstants.java
… src/library/scala/runtime/ScalaRunTime.scala
… src/library/scala/runtime/StructuralCallSite.java
… src/library/scala/runtime/StructuralCallSite.scala
… src/library/scala/runtime/java8/JFunction0$mcB$sp.java
… src/library/scala/runtime/java8/JFunction0$mcB$sp.scala
… src/library/scala/runtime/java8/JFunction0$mcC$sp.java
… src/library/scala/runtime/java8/JFunction0$mcC$sp.scala
… src/library/scala/runtime/java8/JFunction0$mcD$sp.java
… src/library/scala/runtime/java8/JFunction0$mcD$sp.scala