@lrytz 2018-08-24
Test failure due to more inlining, leading to https://github.com/scala/bug/issues/9137. Oh and travis doesn't like the long time spent in the inliner :) `No output has been received in the last 10m0s`.