@lrytz 2018-09-27
This is finally getting ready now! I summarized the changes in the [PR description](https://github.com/scala/scala/pull/7133#issue-210695442). Feedback is very welcome, /cc @Ichoran, @non, @mkeskells, @fwbrasil