@lrytz 2018-09-28
Pushed changes (proposed and written by @retronym) to cut dependencies of Java sources on Scala files in the library, and switched to JavaThenScala. This allows inlining code that references those Java-defined types, see https://github.com/scala/scala/pull/6085#issuecomment-356827948. This shows a slight performance improvement, see 3rd data piont in https://scala-ci.typesafe.com/grafana/dashboard/db/pr-validation?orgId=1&var-scalaSha=84de40d%7C9eef228%7C8d17277&var-source=All&var-bench=HotScalacBenchmark.compile&var-host=scalabench@scalabench@&from=1535327070397&to=1538146460955