@lrytz 2018-10-05
I ran the compiler benchmark `hot -psource=scalap -f 1 -wi 20 -w 10 -i 10 -r 10` for 3 scala versions

  - baseline compiled with the optimizer disabled
  - baseline compiled with `-opt`
  - this branch compiled with `-opt`

on 3 vms

  - Oracle 1.8.0_181 / HotSpot
  - Graal CE 1.0 RC7
  - Graal EE 1.0 RC7

on 2 machines

  - Our benchmarking machine (debian)
  - My MacBook Pro 2018


Results on the benchmarking machine (unit is `ms/op`). In parentheses is the factor to the baseline.

| | 2.13.0-pre-84de40d-noopt | 2.13.0-pre-84de40d | 2.13.0-pre-8d17277-SNAPSHOT |
| --- | --- | --- | --- |
| HotSpot | 742.931 ± 2 (1.0) | 713.849 ± 2 (0.961) | 678.911 ± 2 (0.914) |
| Graal CE | 695.825 ± 4 (0.937) | 662.258 ± 2 (0.891) | 660.256 ± 2 (0.889) |
| Graal EE | 578.633 ± 3 (0.779) | 565.439 ± 2 (0.761) | 552.111 ± 2 (0.743) |

Summary:
  - The optimzier helps on HotSpot (makes the compiler ~9% faster)
  - For Graal CE, the optimizer is still useful (makes the compiler ~5% faster)
  - Graal EE helps a lot. The non-optimzied compiler on Graal EE runs 22% faster than on HotSpot, and 15% faster than the optimized compiler on HotSpot.
  - The optimzier doesn't help much on Graal EE, but it doesn't hurt.

Results on my MacBook Pro

| | 2.13.0-pre-84de40d-noopt | 2.13.0-pre-84de40d | 2.13.0-pre-8d17277-SNAPSHOT |
| --- | --- | --- | --- |
| HotSpot | 743.920 ± 7 (1.00) | 710.432 ± 6 (0.955) | 694.129 ± 6 (0.933) |
| Graal CE | 840.872 ± 9 (1.130) | 813.020 ± 8 (1.093) | 803.120 ± 7 (1.080) |
| Graal EE | 741.984 ± 7 (0.997) | 723.949 ± 8 (0.973) | 711.624 ± 6 (0.957) |

I have no idea why Graal doesn't do better on my mac. The "VM options" in the JMH log are the same
```
[info] # VM options: -Xms2G -Xmx2G -Xss2M
```
Maybe it has to do with thermal throttling on these modern laptops? I'll reach out to some other people to run the same benchmarks.