@SethTisue 2018-10-09
community build run queued: 
https://scala-ci.typesafe.com/job/scala-2.13.x-integrate-community-build/1493/

currently 53 projects are expected to be green, so it gives some assurance, not as much as the 2.12.x community build of course