@edmundnoble 2018-10-14
Love these changes, I can see them making a lot of difference especially to heavy collections users, and I think they'll make people more confident in their understanding of what can and can't be inlined.