@retronym 2018-10-16
@lrytz Here's a commit to update the `ScalaOptionParser` in our SBT build: https://github.com/scala/scala/compare/2.13.x...retronym:review/7133?expand=1