@mkeskells 2018-10-16
@lrytz Impressive! Looks like the work to improve expanded to be a rewrite!
I am very late to the party here and will make some time to have a look in more detail this week
I have a few questions in the interim though

- has this been tested with the parallel backend - if nothing else this should be able to reduce some of the extra overheads of inliner as inliner is part of the for local optimisations isn't it?

- is this change readily back-portable to 2.12 - I know there are 200+ files changed but a lot of non jvm stuff that looks mechanical (FunctionN etc), and I thought that jvm was quite similar

- did you already look at the inliner warnings in the current build - I presume this will be a separate PR(e.g. a PR - better use the new inliner), but I recall there were a bunch of inliner warnings that I used to see in making a new dist