@mdedetrich 2018-10-26
> Technically I don't see a problem backporting it to 2.12. So it's a question of risk, it's a big change that impacts generated bytecode. We should at least gain a bit of confidence first and see how well it does on 2.13.x.

Doesn't this effectively rule out a change for Scala 2.12.x?