@lrytz 2018-10-26
> Doesn't this effectively rule out a change for Scala 2.12.x?

I don't think that it's ruled out by any principles. Binary compatible bytecode changes are OK, so it's a risk-value tradeoff.